Clean up comments and documentation, restruct project structure for navigation, add bottom bar navigation via expo-router
