+++ client/package.json
@@ -16,7 +16,7 @@
   },
   "dependencies": {
     "@expo/metro-runtime": "~4.0.1",
-    "@expo/vector-icons": "^14.0.2",
+    "@expo/vector-icons": "^14.0.4",
     "@react-native-async-storage/async-storage": "^2.1.1",
     "@react-native-google-signin/google-signin": "^13.2.0",
     "@react-navigation/bottom-tabs": "^7.2.0",
@@ -58,7 +58,7 @@
     "react-native-gesture-handler": "~2.20.2",
     "react-native-permissions": "^5.2.5",
     "react-native-reanimated": "~3.16.1",
-    "react-native-safe-area-context": "4.12.0",
+    "react-native-safe-area-context": "^4.12.0",
     "react-native-screens": "~4.4.0",
     "react-native-web": "~0.19.13",
     "react-native-webview": "13.12.5",
