+++ client/src/screens/(tabs)/_layout.tsx
@@ -1,0 +1,81 @@
+import React from 'react';
+import { Tabs } from "expo-router";
+import { View, TouchableOpacity, StyleSheet } from "react-native";
+import { Ionicons } from "@expo/vector-icons";
+import { useSafeAreaInsets } from "react-native-safe-area-context";
+
+export default function TabsLayout() {
+  const insets = useSafeAreaInsets();
+
+  const CustomTabBarButton = ({ children, onPress }) => (
+    <TouchableOpacity
+      style={{
+        top: -20,
+        justifyContent: 'center',
+        alignItems: 'center',
+      }}
+      onPress={onPress}
+    >
+      <View style={{
+        width: 70,
+        height: 70,
+        borderRadius: 35,
+        backgroundColor: '#00cc99',
+        ...styles.shadow
+      }}>
+        {children}
+      </View>
+    </TouchableOpacity>
+  );
+
+  return (
+    <Tabs
+      screenOptions={{
+        headerShown: false,
+        tabBarShowLabel: false,
+        tabBarStyle: {
+          position: 'absolute',
+          // bottom: insets.bottom,
+          elevation: 0,
+          backgroundColor: '#ffffff',
+          borderTopWidth: 0,
+          height: 60 + (insets.bottom > 0 ? insets.bottom - 5 : 0),
+          ...styles.shadow
+        },
+      }}
+    >
+      <Tabs.Screen
+        name="index"
+        options={{
+          title: "Home",
+          tabBarIcon: ({ focused }) => (
+            <Ionicons name="home" size={24} color={focused ? "#00cc99" : "#748c94"} />
+          ),
+        }}
+      />
+
+      <Tabs.Screen
+        name="events"
+        options={{
+          title: "Events",
+          tabBarIcon: ({ focused }) => (
+            <Ionicons name="calendar" size={24} color={focused ? "#00cc99" : "#748c94"} />
+          ),
+        }}
+      />
+    </Tabs>
+  );
+}
+
+const styles = StyleSheet.create({
+  shadow: {
+    shadowColor: '#7F5DF0',
+    shadowOffset: {
+      width: 0,
+      height: 10,
+    },
+    shadowOpacity: 0.25,
+    shadowRadius: 3.5,
+    elevation: 5
+  }
+});
