+++ client/src/screens/_layout.tsx
@@ -6,7 +6,5 @@ export default function RootLayout() {
     <Stack.Screen name="register" options={{ title: 'Register' }} />
     <Stack.Screen name="login" options={{ title: 'Sign in' }} />
     <Stack.Screen name="home" options={{ title: 'Home' }} />
-    <Stack.Screen name="events" options={{ gestureEnabled: true }} />
   </Stack>;
-  // return <RegisterScreen />;
 }
